# Conflicts:
#	JukeboxMC-Server/src/main/kotlin/org/jukeboxmc/server/block/behavior/BlockCoralFan.kt
#	JukeboxMC-Server/src/main/kotlin/org/jukeboxmc/server/block/behavior/BlockCoralFanDead.kt
#	JukeboxMC-Server/src/main/kotlin/org/jukeboxmc/server/block/behavior/BlockRedFlower.kt
#	JukeboxMC-Server/src/main/kotlin/org/jukeboxmc/server/item/ItemRegistry.kt

+++ JukeboxMC-Server/src/main/kotlin/org/jukeboxmc/server/block/behavior/BlockSoulLantern.kt
@@ -6,14 +6,16 @@ import org.jukeboxmc.api.block.SoulLantern
 import org.jukeboxmc.server.block.JukeboxBlock
 import org.jukeboxmc.server.extensions.toByte
 
-class BlockSoulLantern(identifier: Identifier, blockStates: NbtMap?) : JukeboxBlock(identifier, blockStates), SoulLantern {
+class BlockSoulLantern(identifier: Identifier, blockStates: NbtMap?) : JukeboxBlock(identifier, blockStates),
+    SoulLantern {
 
-   override fun isHanging(): Boolean {
-       return this.getBooleanState("hanging")
-   }
+    override fun isHanging(): Boolean {
+        return this.getBooleanState("hanging")
+    }
 
-   override fun setHanging(value: Boolean): SoulLantern {
-       return this.setState("hanging", value.
-       toByte())
-   }
+    override fun setHanging(value: Boolean): SoulLantern {
+        return this.setState(
+            "hanging", value.toByte()
+        )
+    }
 }
